Remove "char.no-pad" flag
Given the existing implementation in presto/master there is no way of correctly implementing no-padding version.
Whenever there is a length mismatch, coercion is triggered before comparison occurs.
Padding is inevitable for chars of the same declared length because we don't store spaces, thus there is no way to know the actual value of char with spaces.
E.g. when user inserted 'a ' into CHAR(3) it's stored as 'a' and length = 3. It's unknown whether user typed 'a', 'a ' or 'a  ' though.