Replace distributed_join w/ join_distribution_type
Replace distributed-joins-enabled and distributed_join with a
join_distribution_type property. This property includes three options:
"partitioned", "replicated", and "automatic". "Partitioned" has the same
behavior as distribtued_joins=true. "Replicated" is equivalent to
distributed_joins=false. "Automatic" will use stats to evaluate whether
partitioned or replicated is better, or if no information is available,
it will choose partitioned. We need to entirely remove the old
distributed_joins property because it conflicts with the new one, and so
the behavior would be incorrect.